src/requirements.txt
awscli==1.34.5
black==24.8.0
boto3==1.35.5
botocore==1.35.5
click==8.1.7
colorama==0.4.6
docutils==0.16
isort==5.13.2
jmespath==1.0.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pdf2image==1.17.0
pillow==10.4.0
platformdirs==4.2.2
pyasn1==0.6.0
pycodestyle==2.12.1
pydocstyle==6.3.0
pyflakes==3.2.0
pylama==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
rsa==4.7.2
s3transfer==0.10.2
setuptools==74.1.2
six==1.16.0
snowballstemmer==2.2.0
urllib3==2.2.2
